Use separate types/ files for the main documented FidesJS interfaces
