+++ clients/fides-js/src/lib/events.ts
@@ -1,20 +1,6 @@
 import { debugLog } from "./consent-utils";
 import { FidesCookie } from "./consent-types";
-
-/**
- * Defines the available event names:
- * - FidesInitialized: dispatched when initialization is complete, from Fides.init()
- * - FidesUpdated: dispatched when preferences are updated from updateConsentPreferences()
- * - FidesUIShown: dispatched when either the banner or modal is shown to the user
- * - FidesUIChanged: dispatched when preferences are changed but not saved, i.e. "dirty".
- * - FidesModalClosed: dispatched when the modal is closed
- */
-export type FidesEventType =
-  | "FidesInitialized"
-  | "FidesUpdated"
-  | "FidesUIShown"
-  | "FidesUIChanged"
-  | "FidesModalClosed";
+import type { FidesEventType } from "../types";
 
 // Bonus points: update the WindowEventMap interface with our custom event types
 declare global {
@@ -44,6 +30,11 @@ export type FidesEventDetail = FidesCookie & {
   extraDetails?: FidesEventExtraDetails;
 };
 
+/**
+ * TODO (PROD-1815): Replace this type with this: import { FidesEvent } from "../types"
+ * 
+ * However, this will require locking down some types and refactoring usage.
+ */
 export type FidesEvent = CustomEvent<FidesEventDetail>;
 
 /**
